+++ src/examples/wide-angle-camera/main.cpp
@@ -22,7 +22,7 @@ raytracer::ref<raytracer::Hittable> createWorld()
 int main(int argc, char** argv)
 {
 	constexpr int width = 400;
-	constexpr float aspcet_ratio = 16.0/9.0;
+	constexpr double aspcet_ratio = 16.0/9.0;
 
 	// World
 	auto world = createWorld();
