+++ src/raytracer/materials/dielectric.h
@@ -8,17 +8,17 @@ namespace raytracer
 	class Dielectric : public Material
 	{
 	public:
-		explicit Dielectric(float index_of_refraction);
+		explicit Dielectric(double index_of_refraction);
 
 		~Dielectric() override = default;
 
 		bool scatter(const Ray& r_in, const hit_record& rec, Color& attenuation, Ray& scattered) const override;
 
 	private:
-		static float reflectance(float cosine, float ref_idx);
+		static double reflectance(double cosine, double ref_idx);
 
 	private:
-		float m_ir;
+		double m_ir;
 	};
 
 } // raytracer
