+++ src/raytracer/textures/texture.h
@@ -10,7 +10,7 @@ namespace raytracer
 	public:
 		virtual ~Texture() = default;
 
-		[[nodiscard]] virtual Color value(float u, float v, const Point3& p) const = 0;
+		[[nodiscard]] virtual Color value(double u, double v, const Point3& p) const = 0;
 	};
 
 } // raytracer
